LBJ-208 restructuring the packages to a semantic(eventlog, snapshots, transmissions) instead of a layered (service, web, persistence, ...) structure. Thereby splitting up and inlining the EventLogMapper into the classes that previously called it. Snpashot Events will therefore now also be created via the EventLogWriter to keep the EventLog mapping logic in one place. This commit therefore removes the batching from the snapshot creation, it is planned to be recreated by providing pagination in the snapshot creation in one of the next commits.
