+++ src/main/java/org/zalando/nakadiproducer/NakadiProducerAutoConfiguration.java
@@ -16,7 +16,6 @@ import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 import org.zalando.fahrschein.NakadiClient;
 import org.zalando.nakadiproducer.snapshots.SnapshotEventProvider;
 import org.zalando.nakadiproducer.snapshots.SnapshotEventProviderNotImplementedException;
-import org.zalando.nakadiproducer.web.FlowIdComponent;
 import org.zalando.tracer.Tracer;
 
 @Configuration
