+++ src/main/java/org/zalando/nakadiproducer/eventlog/impl/EventDataOperation.java
@@ -1,0 +1,23 @@
+package org.zalando.nakadiproducer.eventlog.impl;
+
+public enum EventDataOperation {
+
+    CREATE("C"),
+
+    UPDATE("U"),
+
+    DELETE("D"),
+
+    SNAPSHOT("S");
+
+    private final String value;
+
+    EventDataOperation(final String value) {
+        this.value = value;
+    }
+
+    @Override
+    public String toString() {
+        return value;
+    }
+}
