--- src/main/java/org/zalando/nakadiproducer/persistence/entity/EventDataOperation.java
@@ -1,23 +1,0 @@
-package org.zalando.nakadiproducer.persistence.entity;
-
-public enum EventDataOperation {
-
-    CREATE("C"),
-
-    UPDATE("U"),
-
-    DELETE("D"),
-
-    SNAPSHOT("S");
-
-    private final String value;
-
-    EventDataOperation(final String value) {
-        this.value = value;
-    }
-
-    @Override
-    public String toString() {
-        return value;
-    }
-}
