--- src/main/java/org/zalando/nakadiproducer/persistence/entity/EventLog.java
@@ -1,58 +1,0 @@
-package org.zalando.nakadiproducer.persistence.entity;
-
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.Instant;
-
-import javax.persistence.Entity;
-import javax.persistence.EntityListeners;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
-
-@Entity
-@EntityListeners(AuditingEntityListener.class)
-@Table(schema = "nakadi_events", name = "event_log")
-@ToString
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder(toBuilder = true)
-public class EventLog {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-
-    private String eventType;
-
-    private String eventBodyData;
-
-    private String dataType;
-
-    private String dataOp;
-
-    private String flowId;
-
-    @CreatedDate
-    private Instant created;
-
-    @LastModifiedDate
-    private Instant lastModified;
-
-    private String lockedBy;
-
-    private Instant lockedUntil;
-
-}
