+++ src/main/java/org/zalando/nakadiproducer/transmission/EventTransmitter.java
@@ -1,0 +1,5 @@
+package org.zalando.nakadiproducer.transmission;
+
+public interface EventTransmitter {
+    void sendEvents();
+}
