+++ src/main/java/org/zalando/nakadiproducer/transmission/impl/NakadiMetadata.java
@@ -1,0 +1,20 @@
+package org.zalando.nakadiproducer.transmission.impl;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+import java.time.Instant;
+
+import lombok.Data;
+
+@Data
+public class NakadiMetadata {
+
+    @JsonProperty("eid")
+    private String eid;
+
+    @JsonProperty("occurred_at")
+    @JsonFormat(shape = JsonFormat.Shape.STRING)
+    private Instant occuredAt;
+
+}
